refactor(postgres): split Checkpointer in separate component
