+++ eventstore/postgres/go.mod
@@ -3,8 +3,25 @@ module github.com/get-eventually/go-eventually/eventstore/postgres
 go 1.15
 
 require (
-	github.com/get-eventually/go-eventually v0.0.0-20210513211936-20a5f17a39fa
+	github.com/Microsoft/go-winio v0.4.16 // indirect
+	github.com/containerd/containerd v1.4.3 // indirect
+	github.com/docker/distribution v2.7.1+incompatible // indirect
+	github.com/docker/docker v20.10.2+incompatible // indirect
+	github.com/docker/go-connections v0.4.0 // indirect
+	github.com/docker/go-units v0.4.0 // indirect
+	github.com/get-eventually/go-eventually v0.0.0-20210823224756-746a3204cbea
+	github.com/gogo/protobuf v1.3.2 // indirect
 	github.com/golang-migrate/migrate v3.5.4+incompatible
+	github.com/gorilla/mux v1.8.0 // indirect
 	github.com/lib/pq v1.10.1
+	github.com/moby/term v0.0.0-20201216013528-df9cb8a40635 // indirect
+	github.com/morikuni/aec v1.0.0 // indirect
+	github.com/opencontainers/go-digest v1.0.0 // indirect
+	github.com/opencontainers/image-spec v1.0.1 // indirect
+	github.com/sirupsen/logrus v1.7.0 // indirect
 	github.com/stretchr/testify v1.7.0
+	go.uber.org/zap v1.18.1
+	golang.org/x/time v0.0.0-20201208040808-7e3f01d25324 // indirect
+	google.golang.org/grpc v1.34.1 // indirect
+	gotest.tools/v3 v3.0.3 // indirect
 )
